Update map.adoc
In the original code example, it uses the word "val" to represent too many different things, which is confusing for people who don't already understand how the map function works.  The example shows an integer named "val" being created, but then uses "val" throughout the entire function.  This leads students to assume that the integer "val" is supposed to be used 4 times, but that isn't the case.  There should be something descriptive like what I added.  The name of the integers needs to A) be descriptive of what they represent, and B) be uniquely named that that people understand that there is more than 1 integer in the code.